+++ src/install/init/systemd.rs
@@ -4,11 +4,11 @@
 // public it makes no sense to document errors.
 
 use std::ffi::OsStr;
-use std::io;
 use std::path::{Component, Path, PathBuf};
 use std::process::Command;
 use std::thread;
 use std::time::{Duration, Instant};
+use std::{fs, io};
 
 use crate::install::builder::Trigger;
 use crate::install::files::NoHomeError;
@@ -28,8 +28,8 @@ pub use disable_existing::DisableError;
 
 #[derive(thiserror::Error, Debug)]
 pub enum SystemCtlError {
-    #[error("Could not run systemctl, error: {0}")]
-    Io(#[from] std::io::Error),
+    #[error("Could not run systemctl")]
+    Io(#[from] #[source] std::io::Error),
     #[error("Systemctl failed: {reason}")]
     Failed { reason: String },
     #[error("Timed out trying to enable service")]
@@ -45,15 +45,31 @@ pub enum SystemCtlError {
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
     #[error("Could not configure systemd")]
-    SystemCtl(#[from] #[source] SystemCtlError),
+    SystemCtl(
+        #[from]
+        #[source]
+        SystemCtlError,
+    ),
     #[error("Could not write out unit file to {path}")]
-    Writing { #[source] e: io::Error, path: PathBuf },
+    Writing {
+        #[source]
+        e: io::Error,
+        path: PathBuf,
+    },
     #[error("Could not remove the unit files, error: {0}")]
     Removing(#[source] io::Error),
     #[error("Could not verify unit files where created by us, could not open them")]
-    Verifying(#[from] #[source] unit::Error),
+    Verifying(
+        #[from]
+        #[source]
+        unit::Error,
+    ),
     #[error("Could not check if this system uses systemd")]
-    CheckingInitSys(#[from] #[source] PathCheckError),
+    CheckingInitSys(
+        #[from]
+        #[source]
+        PathCheckError,
+    ),
 }
 
 pub(crate) fn path_is_systemd(path: &Path) -> Result<bool, PathCheckError> {
@@ -69,7 +85,7 @@ pub(crate) fn path_is_systemd(path: &Path) -> Result<bool, PathCheckError> {
         .any(|c| c == "systemd"))
 }
 
-// check if systemd is the init system (pid 1)
+// check if systemd is the init system (PID 1)
 pub(super) fn not_available() -> Result<bool, SetupError> {
     use sysinfo::{Pid, ProcessRefreshKind, System, UpdateKind};
     let mut s = System::new();
@@ -99,45 +115,59 @@ pub(super) fn set_up_steps(params: &Params) -> Result<Steps, SetupError> {
     })
 }
 
-pub(super) fn tear_down_steps(
-    name: &str,
-    mode: Mode,
-) -> Result<Option<(RSteps, ExeLocation)>, TearDownError> {
-    let without_extension = match mode {
+pub(super) fn tear_down_steps(mode: Mode) -> Result<Option<(RSteps, ExeLocation)>, TearDownError> {
+    let dir = match mode {
         Mode::User => user_path()?,
         Mode::System => system_path(),
-    }
-    .join(name);
-
-    let mut steps = Vec::new();
-
-    let timer_path = without_extension.with_extension("timer");
-    let timer = Unit::from_path(timer_path).map_err(Error::Verifying)?;
-    if timer.our_service() {
-        steps.extend(teardown::disable_then_remove_with_timer(
-            timer.path.clone(),
-            name,
-            mode,
-        ));
-    }
-
-    let service_path = without_extension.with_extension("service");
-    let service = Unit::from_path(service_path).map_err(Error::Verifying)?;
-
-    let exe_path = if service.our_service() {
-        steps.extend(teardown::disable_then_remove_service(
-            service.path.clone(),
-            name,
-            mode,
-        ));
-        service.exe_path().map_err(TearDownError::FindingExePath)?
-    } else if timer.our_service() {
-        return Err(TearDownError::TimerWithoutService);
-    } else {
-        return Ok(None);
     };
 
-    Ok(Some((steps, exe_path)))
+    let mut steps = Vec::new();
+    let mut exe_paths = Vec::new();
+
+    for entry in fs::read_dir(dir).unwrap() {
+        let path = entry.unwrap().path();
+        if path.is_dir() {
+            continue;
+        }
+        let Some(extension) = path.extension().and_then(OsStr::to_str) else {
+            continue;
+        };
+        let unit = Unit::from_path(path.clone()).unwrap();
+        if !unit.our_service() {
+            continue;
+        }
+        let Some(service_name) = path.file_stem().and_then(OsStr::to_str) else {
+            continue;
+        };
+
+        match extension {
+            "timer" => {
+                steps.extend(teardown::disable_then_remove_with_timer(
+                    unit.path.clone(),
+                    service_name,
+                    mode,
+                ));
+            }
+            "service" => {
+                steps.extend(teardown::disable_then_remove_service(
+                    unit.path.clone(),
+                    service_name,
+                    mode,
+                ));
+                exe_paths.push(unit.exe_path().map_err(TearDownError::FindingExePath)?);
+            }
+            _ => continue,
+        }
+    }
+
+    exe_paths.dedup();
+    match (steps.len(), exe_paths.as_slice()) {
+        (0, []) => Ok(None),
+        (0, [_, ..]) => unreachable!("if we get an exe path we got one service to remove"),
+        (1.., []) => Err(TearDownError::TimerWithoutService),
+        (1.., [exe_path]) => Ok(Some((steps, exe_path.clone()))),
+        (1.., _) => Err(TearDownError::MultipleExePaths(exe_paths)),
+    }
 }
 
 /// There are other paths, but for now we return the most commonly used one