+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   display the error chain.
 ### Fixed
 - rollback was performed in the wrong order
+- `prepare_remove` no longer crashes when service name was not provided
 
 # Version 0.5.2 (2024-12-22)
 ### Fixed